testArcgisSDK01/arcgis-flood-routing-demo/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
shapely==2.0.6
pytest==8.3.3
httpx==0.27.2
